third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-simple-content-overflow-expected.txt

It has a visible link_1.










... and an overflowed link like this.








This link should NOT get focused.

PASS successfullyParsed is true

TEST COMPLETE
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
